Extract SpecialAccessAvailability
